Remove CommandTicket to avoid Vec allocation
